refactor(simulation): split page state and view
